Remove the crap out the modernizr module
Modernizr module uses the domReady event for loading it via a script
tag... not cool. And it requires you to put the file in the module
directory rather than the sites/all/libraries directory. Also not cool.
I thought there maybe had been discovered the "Drupal way" of including
Modernizr, but I'll just continue to include it manually sans-module.