Remove LABjs and theme skeleton
LABjs because it doesn't play nice with being installed inside of an
installation profile... or views, really. It seems not quite ready for
prime-time yet.

Skeleton theme because it doesn't belong here. We don't want it to
persist, like it would every time we run the make file. It will be moved
into the shovel script.